RT #61305: return { foo => 1 } should parse curlys as hash constructor, not block
The perlref documentation says that placing "+" or "return" before curly
brackets causes them to not to be interpreted as a block. The examples
seem to make clear that this is a roundabout way of saying that the
curly brackets in "return { foo => 1 }" are always interpreted as a hash
constructor.

This commit attempts to make this happen. It is coded along the same
lines as SVN commit 8963, which made a leading "+" force interpretation
as a hash constructor, and in fact given that code the fix is a
one-liner, though the lexer tests were updated as well.